Locks: Rework to use Update and enforce status changes
When a task run, it shouldn't be possible to return from a Task with the
condition status still marked as Locked. It would cause havoc to the
user and generally, the Lock should always be released when the
condition return. Instead of guessing what the status should be at
the end (or worse, have something like `Unlock` that would mean
nothing), it's better to mark the Condition as errored and let the user
fix their mistake.