Lock: Return Task error back to the user
The Execute() method would only return the final error value from the
last update call to the K8s API Server. This means that errors from the
task would be swallowed from the stack and would only appear in the
conditions itself.

The problem with this strategy is that it's possible a user of the Lock
wants to log the error from the Task. Now they would need to do it from
within the Task which would impact the dev experience.

Instead, the Execute method should only return the most recent error
occurred. This means that under normal k8s API server usage, the task
error will be returned.

However if the API server was to return an error, it would take
precedence and would be the one returned.

It's possible at some point to use errors wrapping, but I feel that the
API for wrapping errors isn't that great in this scenario. Until I have
a clearer idea of the pros/cons, returning the most recent error seems
like an acceptable tradeoff.